Refatoração parcial do código - controller, business e router
